% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_attachment_to_active_conversation}
\alias{add_attachment_to_active_conversation}
\title{Add attachment to active conversation}
\usage{
add_attachment_to_active_conversation(name, content)
}
\arguments{
\item{name}{File name.}

\item{content}{File content as string.}
}
\value{
Logical.
}
\description{
Add attachment to active conversation
}
