% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{available_openai_models}
\alias{available_openai_models}
\title{List of available AI models for selection in the UI}
\usage{
available_openai_models
}
\description{
List of available AI models for selection in the UI
}
