% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_addin.R
\name{chat_addin}
\alias{chat_addin}
\title{Uruchomienie gadżetu czatu LLM z kartami konwersacji}
\usage{
chat_addin()
}
\description{
Główna funkcja uruchamiająca gadżet Shiny. Pozwala na zarządzanie
konwersacjami w kartach oraz zmianę ustawień zaawansowanych (temperatura,
komunikat systemowy) dla każdej z nich.
}
