% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{create_new_conversation}
\alias{create_new_conversation}
\title{Create a new conversation}
\usage{
create_new_conversation(
  activate = FALSE,
  add_initial_settings = TRUE,
  title = NULL
)
}
\arguments{
\item{activate}{Logical. Activate immediately?}

\item{add_initial_settings}{Logical. Add default model and system message?}

\item{title}{Optional title; if NULL, a time-based title is used.}
}
\value{
Character: conversation ID.
}
\description{
Create a new conversation
}
