% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{delete_conversation}
\alias{delete_conversation}
\title{Delete a conversation}
\usage{
delete_conversation(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
TRUE if deleted, FALSE otherwise.
}
\description{
Delete a conversation
}
