% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{generate_conversation_id}
\alias{generate_conversation_id}
\title{Generowanie unikalnego ID dla konwersacji}
\usage{
generate_conversation_id()
}
\value{
Unikalny identyfikator (string)
}
\description{
Generowanie unikalnego ID dla konwersacji
}
