% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_chat_history}
\alias{get_active_chat_history}
\title{Get active chat history}
\usage{
get_active_chat_history()
}
\value{
List of messages (possibly empty).
}
\description{
Get active chat history
}
