% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation}
\alias{get_active_conversation}
\title{Get active conversation object}
\usage{
get_active_conversation()
}
\value{
List or NULL.
}
\description{
Get active conversation object
}
