% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_attachments}
\alias{get_active_conversation_attachments}
\title{Get attachments for active conversation}
\usage{
get_active_conversation_attachments()
}
\value{
List (possibly empty).
}
\description{
Get attachments for active conversation
}
