% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_id}
\alias{get_active_conversation_id}
\title{Get active conversation ID}
\usage{
get_active_conversation_id()
}
\value{
Character or NULL.
}
\description{
Get active conversation ID
}
