% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_all_conversation_ids}
\alias{get_all_conversation_ids}
\title{Get all conversation IDs}
\usage{
get_all_conversation_ids()
}
\value{
Character vector.
}
\description{
Get all conversation IDs
}
