% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_attachments}
\alias{get_conversation_attachments}
\title{Get attachments by ID}
\usage{
get_conversation_attachments(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List or NULL.
}
\description{
Get attachments by ID
}
