% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_data}
\alias{get_conversation_data}
\title{Get conversation data by ID}
\usage{
get_conversation_data(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List or NULL.
}
\description{
Get conversation data by ID
}
