% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_history}
\alias{get_conversation_history}
\title{Get conversation history by ID}
\usage{
get_conversation_history(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List or NULL.
}
\description{
Get conversation history by ID
}
