% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_list_for_display}
\alias{get_conversation_list_for_display}
\title{Pobieranie listy konwersacji do wyświetlenia}
\usage{
get_conversation_list_for_display()
}
\value{
Lista konwersacji (każda to list(id, title)).
}
\description{
Zwraca listę, gdzie każdy element to lista z \code{id} i \code{title}.
Sortuje wg czasu utworzenia (najnowsze na początku).
}
