% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_model}
\alias{get_conversation_model}
\title{Get model for conversation}
\usage{
get_conversation_model(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
Character or NULL.
}
\description{
Get model for conversation
}
