% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_title}
\alias{get_conversation_title}
\title{Get conversation title by ID}
\usage{
get_conversation_title(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
Character or NULL.
}
\description{
Get conversation title by ID
}
