% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{initialize_history_manager}
\alias{initialize_history_manager}
\title{Initialize the history manager}
\usage{
initialize_history_manager()
}
\value{
Character: ID of the created conversation.
}
\description{
Clears state and creates a single new conversation, then activates it.
}
