% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{is_conversation_started}
\alias{is_conversation_started}
\title{Has the conversation started (model locked)?}
\usage{
is_conversation_started(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
Logical.
}
\description{
Has the conversation started (model locked)?
}
