% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_chat_app.R
\name{packetllm_addin}
\alias{packetllm_addin}
\title{Launch PacketLLM from the RStudio Addins menu}
\usage{
packetllm_addin()
}
\value{
Value passed to \code{run_llm_chat_app()}.
}
\description{
Launch PacketLLM from the RStudio Addins menu
}
