% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{read_file_content}
\alias{read_file_content}
\title{Read file content}
\usage{
read_file_content(file_path, pages = NULL)
}
\arguments{
\item{file_path}{Character string. Path to the file.}

\item{pages}{Optional. A numeric vector specifying which pages (for PDF) should be read.}
}
\value{
A character string containing the file content.
}
\description{
This function reads the content of a file with the extension .R, .pdf, or .docx
and returns it as a single character string.
For PDF files, if the \code{pages} parameter is provided, only the selected pages will be read.
}
