% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_active_conversation}
\alias{set_active_conversation}
\title{Set the active conversation}
\usage{
set_active_conversation(id)
}
\arguments{
\item{id}{Conversation ID or NULL.}
}
\value{
Invisible NULL.
}
\description{
Set the active conversation
}
