% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_model}
\alias{set_conversation_model}
\title{Set model for conversation (if not started)}
\usage{
set_conversation_model(id, model_name)
}
\arguments{
\item{id}{Conversation ID.}

\item{model_name}{Model name (must be in available_openai_models).}
}
\value{
Logical.
}
\description{
Set model for conversation (if not started)
}
