% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_system_message}
\alias{set_conversation_system_message}
\title{Set system message for conversation}
\usage{
set_conversation_system_message(id, message)
}
\arguments{
\item{id}{Conversation ID.}

\item{message}{Single string system message.}
}
\value{
Logical.
}
\description{
Set system message for conversation
}
