% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{update_conversation_title}
\alias{update_conversation_title}
\title{Aktualizacja tytułu konwersacji}
\usage{
update_conversation_title(id, title)
}
\arguments{
\item{id}{ID konwersacji.}

\item{title}{Nowy tytuł.}
}
\value{
Invisible NULL. Zgłasza błąd, jeśli ID nie istnieje.
}
\description{
Aktualizacja tytułu konwersacji
}
