% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_attachment_to_active_conversation}
\alias{add_attachment_to_active_conversation}
\title{Dodaje załącznik do aktywnej konwersacji}
\usage{
add_attachment_to_active_conversation(name, content)
}
\arguments{
\item{name}{Nazwa pliku załącznika.}

\item{content}{Treść pliku jako string.}
}
\value{
TRUE jeśli dodano, FALSE jeśli plik o tej nazwie już istnieje lub brak aktywnej konwersacji.
}
\description{
Dodaje załącznik do aktywnej konwersacji
}
