% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_attachment_to_active_conversation}
\alias{add_attachment_to_active_conversation}
\title{Dodawanie załącznika do aktywnej konwersacji}
\usage{
add_attachment_to_active_conversation(name, content)
}
\arguments{
\item{name}{Nazwa pliku.}

\item{content}{Treść pliku.}
}
\value{
TRUE jeśli dodano, FALSE jeśli plik istnieje.
}
\description{
Dodawanie załącznika do aktywnej konwersacji
}
