% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_attachment_to_active_conversation}
\alias{add_attachment_to_active_conversation}
\title{Adds an attachment to the active conversation}
\usage{
add_attachment_to_active_conversation(name, content)
}
\arguments{
\item{name}{Name of the attachment file.}

\item{content}{File content as a string.}
}
\value{
TRUE if added, FALSE if a file with this name already exists or no active conversation.
}
\description{
Adds an attachment to the active conversation
}
