% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_message_to_active_history}
\alias{add_message_to_active_history}
\title{Adds a message to the active conversation's history}
\usage{
add_message_to_active_history(role, content)
}
\arguments{
\item{role}{Role ("user", "assistant", "system").}

\item{content}{Message content.}
}
\value{
A list indicating the result:
- \code{list(type = "title_set", new_title = "...")} if a new title was set.
- \code{list(type = "assistant_locked_model")} if the model was locked.
- \code{list(type = "message_added")} in other cases of adding a message.
- \code{list(type = "error", message = "...")} in case of an error.
}
\description{
Updates the conversation title based on content if it's the first user message.
Sets the \code{model_locked} flag to TRUE on the first \strong{assistant} message.
}
