% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_message_to_active_history}
\alias{add_message_to_active_history}
\title{Dodawanie wiadomości do historii aktywnej konwersacji}
\usage{
add_message_to_active_history(role, content)
}
\arguments{
\item{role}{Rola ("user", "assistant", "system").}

\item{content}{Treść wiadomości.}
}
\value{
Invisible NULL. Zgłasza błąd, jeśli żadna konwersacja nie jest aktywna.
}
\description{
Dodawanie wiadomości do historii aktywnej konwersacji
}
