% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{add_message_to_active_history}
\alias{add_message_to_active_history}
\title{Dodawanie wiadomości do historii aktywnej konwersacji}
\usage{
add_message_to_active_history(role, content)
}
\arguments{
\item{role}{Rola ("user", "assistant", "system").}

\item{content}{Treść wiadomości.}
}
\value{
Nowy tytuł konwersacji (string) jeśli został zaktualizowany, w przeciwnym razie NULL.
}
\description{
Aktualizuje tytuł, jeśli to pierwsza wiadomość użytkownika.
}
