% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_logic.R
\name{add_user_message}
\alias{add_user_message}
\title{Dodanie wiadomości użytkownika do aktywnej konwersacji}
\usage{
add_user_message(text)
}
\arguments{
\item{text}{Tekst wiadomości użytkownika.}
}
\value{
Invisible NULL
}
\description{
Dodanie wiadomości użytkownika do aktywnej konwersacji
}
