% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api.R
\docType{data}
\name{available_openai_models}
\alias{available_openai_models}
\title{List of available OpenAI models for selection in the UI}
\format{
An object of class \code{character} of length 5.
}
\usage{
available_openai_models
}
\description{
List of available OpenAI models for selection in the UI
}
\keyword{datasets}
