% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api.R
\docType{data}
\name{available_openai_models}
\alias{available_openai_models}
\title{Lista dostępnych modeli OpenAI do wyboru w UI}
\format{
An object of class \code{character} of length 4.
}
\usage{
available_openai_models
}
\description{
Lista dostępnych modeli OpenAI do wyboru w UI
}
\keyword{datasets}
