% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api.R
\name{call_openai_chat}
\alias{call_openai_chat}
\title{Wywołanie API OpenAI}
\usage{
call_openai_chat(messages, model, temperature = 0.5)
}
\arguments{
\item{messages}{Lista wiadomości (każda wiadomość to lista z polami 'role' i 'content').}

\item{model}{Model OpenAI do użycia.}

\item{temperature}{Parametr temperatury (używany tylko dla modeli, które go obsługują).}
}
\value{
Łańcuch znaków zawierający odpowiedź modelu.
}
\description{
Funkcja wysyła historię konwersacji do API OpenAI i zwraca odpowiedź modelu.
Dla modeli z \code{simplified_models_list} NIE wysyła parametru \code{temperature}.
}
