% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{chat_addin}
\alias{chat_addin}
\title{Uruchomienie gadżetu czatu LLM z historią konwersacji}
\usage{
chat_addin()
}
\description{
Otwiera okno dialogowe Shiny Gadget umożliwiające interakcję z LLM.
Pozwala na prowadzenie wielu rozmów w jednej sesji i przełączanie się
między nimi za pomocą przycisku "Historia".
Dołączanie plików za pomocą przycisku '+'. Przycisk "Wyślij"
jest aktywny tylko gdy jest tekst lub załączone pliki.
Zamykanie okna za pomocą przycisku 'X' zamyka gadżet.
}
