% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api.R
\name{check_api_key}
\alias{check_api_key}
\title{Sprawdzenie klucza API}
\usage{
check_api_key()
}
\value{
Zwraca TRUE, jeśli klucz API jest ustawiony; w przeciwnym razie przerywa działanie.
}
\description{
Ta funkcja sprawdza, czy w zmiennych środowiskowych (np. w pliku .Renviron)
znajduje się klucz API przypisany do zmiennej \code{OPENAI_API_KEY}.
Jeśli klucz nie jest ustawiony, funkcja zgłosi błąd i przerwie działanie.
Jeśli klucz jest ustawiony, zwraca TRUE.
}
