% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api.R
\name{check_api_key}
\alias{check_api_key}
\title{Check API Key}
\usage{
check_api_key()
}
\value{
Returns TRUE if the API key is set; otherwise, stops execution.
}
\description{
This function checks if the API key assigned to the \code{OPENAI_API_KEY} variable
exists in the environment variables (e.g., in the .Renviron file).
If the key is not set, the function will raise an error and stop execution.
If the key is set, it returns TRUE.
}
