% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{create_new_conversation}
\alias{create_new_conversation}
\title{Creates a new conversation}
\usage{
create_new_conversation(
  activate = FALSE,
  add_initial_settings = TRUE,
  title = NULL
)
}
\arguments{
\item{activate}{Should the new conversation be set as active (default FALSE).}

\item{add_initial_settings}{Should a default system message be added to the history (default TRUE).}

\item{title}{Initial title of the conversation (default generated based on time).}
}
\value{
ID of the newly created conversation.
}
\description{
Adds a new conversation to the \code{.history_env} environment.
}
