% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{create_new_conversation}
\alias{create_new_conversation}
\title{Tworzenie nowej konwersacji}
\usage{
create_new_conversation(
  activate = FALSE,
  add_initial_settings = TRUE,
  title = NULL
)
}
\arguments{
\item{activate}{Czy ustawić nową konwersację jako aktywną? (Domyślnie FALSE)}

\item{add_initial_settings}{Czy dodać domyślny komunikat systemowy i temperaturę? (Domyślnie TRUE) # Zmieniono nazwę flagi}

\item{title}{Początkowy tytuł konwersacji (Domyślnie "Nowa rozmowa \link{czas}")}
}
\value{
ID nowo utworzonej konwersacji.
}
\description{
Tworzy nowy obiekt konwersacji z domyślnymi ustawieniami zaawansowanymi.
}
