% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{create_new_conversation}
\alias{create_new_conversation}
\title{Tworzenie nowej konwersacji}
\usage{
create_new_conversation(
  activate = TRUE,
  add_initial_system_msg = TRUE,
  title = NULL
)
}
\arguments{
\item{activate}{Czy ustawić nową konwersację jako aktywną? (Domyślnie TRUE)}

\item{add_initial_system_msg}{Czy dodać domyślny komunikat systemowy? (Domyślnie TRUE)}

\item{title}{Początkowy tytuł konwersacji (Domyślnie "Nowa rozmowa \link{czas}")}
}
\value{
ID nowo utworzonej konwersacji.
}
\description{
Tworzy nowy obiekt konwersacji, dodaje go do listy i opcjonalnie ustawia jako aktywny.
}
