% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{create_new_conversation}
\alias{create_new_conversation}
\title{Tworzy nową konwersację}
\usage{
create_new_conversation(
  activate = FALSE,
  add_initial_settings = TRUE,
  title = NULL
)
}
\arguments{
\item{activate}{Czy ustawić nową konwersację jako aktywną (domyślnie FALSE).}

\item{add_initial_settings}{Czy dodać domyślny komunikat systemowy do historii (domyślnie TRUE).}

\item{title}{Początkowy tytuł konwersacji (domyślnie generowany na podstawie czasu).}
}
\value{
ID nowo utworzonej konwersacji.
}
\description{
Dodaje nową konwersację do środowiska \code{.history_env}.
}
