% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{delete_conversation}
\alias{delete_conversation}
\title{Usuwa konwersację o podanym ID}
\usage{
delete_conversation(id)
}
\arguments{
\item{id}{ID konwersacji do usunięcia.}
}
\value{
TRUE jeśli usunięto, FALSE jeśli nie istniała.
}
\description{
Usuwa konwersację o podanym ID
}
