% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{delete_conversation}
\alias{delete_conversation}
\title{Deletes the conversation with the given ID}
\usage{
delete_conversation(id)
}
\arguments{
\item{id}{ID of the conversation to delete.}
}
\value{
TRUE if deleted, FALSE if it did not exist.
}
\description{
Deletes the conversation with the given ID
}
