% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_chat_history}
\alias{get_active_chat_history}
\title{Pobieranie historii czatu dla aktywnej konwersacji}
\usage{
get_active_chat_history()
}
\value{
Lista wiadomości ({role, content}) dla aktywnej konwersacji lub pusta lista.
}
\description{
Pobieranie historii czatu dla aktywnej konwersacji
}
