% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_chat_history}
\alias{get_active_chat_history}
\title{Pobiera historię czatu dla aktywnej konwersacji}
\usage{
get_active_chat_history()
}
\value{
Lista wiadomości lub pusta lista.
}
\description{
Pobiera historię czatu dla aktywnej konwersacji
}
