% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_chat_history}
\alias{get_active_chat_history}
\title{Gets the chat history for the active conversation}
\usage{
get_active_chat_history()
}
\value{
List of messages or an empty list.
}
\description{
Gets the chat history for the active conversation
}
