% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation}
\alias{get_active_conversation}
\title{Pobiera pełny obiekt aktywnej konwersacji}
\usage{
get_active_conversation()
}
\value{
Lista reprezentująca konwersację lub NULL.
}
\description{
Pobiera pełny obiekt aktywnej konwersacji
}
