% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation}
\alias{get_active_conversation}
\title{Gets the full object of the active conversation}
\usage{
get_active_conversation()
}
\value{
List representing the conversation or NULL.
}
\description{
Gets the full object of the active conversation
}
