% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation}
\alias{get_active_conversation}
\title{Pobieranie obiektu aktywnej konwersacji}
\usage{
get_active_conversation()
}
\value{
Pełny obiekt (lista) aktywnej konwersacji lub NULL.
}
\description{
Pobieranie obiektu aktywnej konwersacji
}
