% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_attachments}
\alias{get_active_conversation_attachments}
\title{Pobieranie załączników dla aktywnej konwersacji}
\usage{
get_active_conversation_attachments()
}
\value{
Lista załączników.
}
\description{
Pobieranie załączników dla aktywnej konwersacji
}
