% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_attachments}
\alias{get_active_conversation_attachments}
\title{Gets the list of attachments for the active conversation}
\usage{
get_active_conversation_attachments()
}
\value{
List of attachments or an empty list.
}
\description{
Gets the list of attachments for the active conversation
}
