% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_id}
\alias{get_active_conversation_id}
\title{Gets the ID of the active conversation}
\usage{
get_active_conversation_id()
}
\value{
ID of the active conversation (string) or NULL.
}
\description{
Gets the ID of the active conversation
}
