% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_active_conversation_id}
\alias{get_active_conversation_id}
\title{Pobieranie ID aktywnej konwersacji}
\usage{
get_active_conversation_id()
}
\value{
ID aktywnej konwersacji lub NULL, jeśli żadna nie jest aktywna.
}
\description{
Pobieranie ID aktywnej konwersacji
}
