% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_all_conversation_ids}
\alias{get_all_conversation_ids}
\title{Gets a list of IDs of all existing conversations}
\usage{
get_all_conversation_ids()
}
\value{
Character vector with conversation IDs.
}
\description{
Gets a list of IDs of all existing conversations
}
