% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_all_conversation_ids}
\alias{get_all_conversation_ids}
\title{Pobieranie listy ID wszystkich konwersacji}
\usage{
get_all_conversation_ids()
}
\value{
Wektor znakowy z ID.
}
\description{
Pobieranie listy ID wszystkich konwersacji
}
