% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_all_conversation_ids}
\alias{get_all_conversation_ids}
\title{Pobiera listę ID wszystkich istniejących konwersacji}
\usage{
get_all_conversation_ids()
}
\value{
Wektor znakowy z ID konwersacji.
}
\description{
Pobiera listę ID wszystkich istniejących konwersacji
}
