% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_attachments}
\alias{get_conversation_attachments}
\title{Pobiera listę załączników dla konwersacji o podanym ID}
\usage{
get_conversation_attachments(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Lista załączników lub NULL, jeśli konwersacja nie istnieje.
}
\description{
Pobiera listę załączników dla konwersacji o podanym ID
}
