% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_attachments}
\alias{get_conversation_attachments}
\title{Gets the list of attachments for the conversation with the given ID}
\usage{
get_conversation_attachments(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List of attachments or NULL if conversation doesn't exist.
}
\description{
Gets the list of attachments for the conversation with the given ID
}
