% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_data}
\alias{get_conversation_data}
\title{Gets the full conversation data object for the given ID}
\usage{
get_conversation_data(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List with conversation data or NULL if it doesn't exist.
}
\description{
Gets the full conversation data object for the given ID
}
