% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_data}
\alias{get_conversation_data}
\title{Pobiera pełny obiekt danych konwersacji o podanym ID}
\usage{
get_conversation_data(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Lista z danymi konwersacji lub NULL, jeśli nie istnieje.
}
\description{
Pobiera pełny obiekt danych konwersacji o podanym ID
}
