% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_history}
\alias{get_conversation_history}
\title{Pobiera historię czatu dla konwersacji o podanym ID}
\usage{
get_conversation_history(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Lista wiadomości lub NULL, jeśli konwersacja nie istnieje.
}
\description{
Pobiera historię czatu dla konwersacji o podanym ID
}
