% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_history}
\alias{get_conversation_history}
\title{Gets the chat history for the conversation with the given ID}
\usage{
get_conversation_history(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
List of messages or NULL if conversation doesn't exist.
}
\description{
Gets the chat history for the conversation with the given ID
}
