% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_model}
\alias{get_conversation_model}
\title{Pobiera nazwę modelu dla konwersacji o podanym ID}
\usage{
get_conversation_model(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Nazwa modelu (string) lub NULL.
}
\description{
Pobiera nazwę modelu dla konwersacji o podanym ID
}
