% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_model}
\alias{get_conversation_model}
\title{Gets the model name for the conversation with the given ID}
\usage{
get_conversation_model(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
Model name (string) or NULL.
}
\description{
Gets the model name for the conversation with the given ID
}
