% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_title}
\alias{get_conversation_title}
\title{Pobiera tytuł konwersacji o podanym ID}
\usage{
get_conversation_title(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Tytuł (string) lub NULL, jeśli konwersacja nie istnieje lub nie ma tytułu.
}
\description{
Pobiera tytuł konwersacji o podanym ID
}
