% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_title}
\alias{get_conversation_title}
\title{Gets the title of the conversation with the given ID}
\usage{
get_conversation_title(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
Title (string) or NULL if the conversation doesn't exist or has no title.
}
\description{
Gets the title of the conversation with the given ID
}
