% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{get_conversation_title}
\alias{get_conversation_title}
\title{Pobieranie tytułu konwersacji}
\usage{
get_conversation_title(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
Tytuł konwersacji lub NULL.
}
\description{
Pobieranie tytułu konwersacji
}
