% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{initialize_history_manager}
\alias{initialize_history_manager}
\title{Initialize the history manager}
\usage{
initialize_history_manager(persist = FALSE)
}
\arguments{
\item{persist}{Logical. Restore and save local gadget history?}
}
\value{
Character: ID of the created conversation.
}
\description{
Initializes state and creates or restores a conversation, then activates it.
}
