% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{initialize_history_manager}
\alias{initialize_history_manager}
\title{Initializes the history manager}
\usage{
initialize_history_manager()
}
\value{
ID of the first created conversation.
}
\description{
Clears the state and creates the first, empty conversation.
}
