% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{initialize_history_manager}
\alias{initialize_history_manager}
\title{Inicjalizacja menedżera historii}
\usage{
initialize_history_manager()
}
\value{
ID nowo utworzonej, pierwszej konwersacji.
}
\description{
Tworzy pierwszą konwersację. Wywoływana na starcie serwera gadżetu.
}
