% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{initialize_history_manager}
\alias{initialize_history_manager}
\title{Inicjalizuje menedżera historii}
\usage{
initialize_history_manager()
}
\value{
ID pierwszej utworzonej konwersacji.
}
\description{
Czyści stan i tworzy pierwszą, pustą konwersację.
}
