% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{is_conversation_started}
\alias{is_conversation_started}
\title{Sprawdza, czy konwersacja została rozpoczęta (model zablokowany)}
\usage{
is_conversation_started(id)
}
\arguments{
\item{id}{ID konwersacji.}
}
\value{
TRUE jeśli model jest zablokowany, FALSE w przeciwnym razie lub jeśli konwersacja nie istnieje.
}
\description{
Sprawdza, czy flaga \code{model_locked} jest TRUE. Model jest blokowany
po dodaniu pierwszej wiadomości \strong{asystenta}.
}
