% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{is_conversation_started}
\alias{is_conversation_started}
\title{Checks if the conversation has started (model locked)}
\usage{
is_conversation_started(id)
}
\arguments{
\item{id}{Conversation ID.}
}
\value{
TRUE if the model is locked, FALSE otherwise or if the conversation doesn't exist.
}
\description{
Checks if the \code{model_locked} flag is TRUE. The model is locked
after the first \strong{assistant} message is added.
}
