% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{parse_pages}
\alias{parse_pages}
\title{Parsowanie zakresu stron}
\usage{
parse_pages(pages_str)
}
\arguments{
\item{pages_str}{Łańcuch znaków określający strony, np. "1-3,5".}
}
\value{
Wektor numeryczny zawierający poszczególne strony.
}
\description{
Funkcja przetwarza ciąg znaków określający zakres stron (np. "1-3,5") i zwraca wektor numeryczny.
}
