% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{read_file_content}
\alias{read_file_content}
\title{Odczyt zawartości pliku}
\usage{
read_file_content(file_path, pages = NULL)
}
\arguments{
\item{file_path}{Łańcuch znaków. Ścieżka do pliku.}

\item{pages}{Opcjonalnie. Wektor numeryczny określający, które strony (dla PDF) mają być odczytane.}
}
\value{
Łańcuch znaków zawierający treść pliku.
}
\description{
Funkcja odczytuje zawartość pliku o rozszerzeniu .R, .pdf lub .docx i zwraca ją jako pojedynczy łańcuch znaków.
W przypadku plików PDF, jeżeli podano parametr \code{pages}, tylko wybrane strony zostaną odczytane.
}
