% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{reset_history_manager}
\alias{reset_history_manager}
\title{Resets the entire state of the history manager}
\usage{
reset_history_manager()
}
\value{
Invisible NULL.
}
\description{
Resets the entire state of the history manager
}
