% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{reset_history_manager}
\alias{reset_history_manager}
\title{Resetowanie całego menedżera historii (na wszelki wypadek)}
\usage{
reset_history_manager()
}
\description{
Resetowanie całego menedżera historii (na wszelki wypadek)
}
