% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{reset_history_manager}
\alias{reset_history_manager}
\title{Resetuje cały stan menedżera historii}
\usage{
reset_history_manager()
}
\value{
Invisible NULL.
}
\description{
Resetuje cały stan menedżera historii
}
