% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_active_conversation}
\alias{set_active_conversation}
\title{Ustawianie aktywnej konwersacji}
\usage{
set_active_conversation(id)
}
\arguments{
\item{id}{ID konwersacji do aktywacji.}
}
\value{
Invisible NULL.
}
\description{
Ustawianie aktywnej konwersacji
}
