% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_active_conversation}
\alias{set_active_conversation}
\title{Sets the active conversation}
\usage{
set_active_conversation(id)
}
\arguments{
\item{id}{ID of the conversation to activate or NULL to deactivate.}
}
\value{
Invisible NULL.
}
\description{
Sets the active conversation
}
