% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_active_conversation}
\alias{set_active_conversation}
\title{Ustawia aktywną konwersację}
\usage{
set_active_conversation(id)
}
\arguments{
\item{id}{ID konwersacji do aktywacji lub NULL, aby zdezaktywować.}
}
\value{
Invisible NULL.
}
\description{
Ustawia aktywną konwersację
}
