% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_model}
\alias{set_conversation_model}
\title{Ustawia model dla konwersacji, jeśli nie została rozpoczęta}
\usage{
set_conversation_model(id, model_name)
}
\arguments{
\item{id}{ID konwersacji.}

\item{model_name}{Nazwa nowego modelu.}
}
\value{
TRUE jeśli model został ustawiony, FALSE w przeciwnym razie.
}
\description{
Sprawdza flagę \code{model_locked} ustawianą po pierwszej odpowiedzi asystenta.
}
