% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_model}
\alias{set_conversation_model}
\title{Sets the model for the conversation, if it hasn't started}
\usage{
set_conversation_model(id, model_name)
}
\arguments{
\item{id}{Conversation ID.}

\item{model_name}{Name of the new model.}
}
\value{
TRUE if the model was set, FALSE otherwise.
}
\description{
Checks the \code{model_locked} flag set after the first assistant response.
}
