% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_system_message}
\alias{set_conversation_system_message}
\title{Ustawia komunikat systemowy dla konwersacji o podanym ID}
\usage{
set_conversation_system_message(id, message)
}
\arguments{
\item{id}{ID konwersacji.}

\item{message}{Nowy komunikat systemowy (string).}
}
\value{
TRUE jeśli ustawiono, FALSE jeśli konwersacja nie istnieje lub wartość jest nieprawidłowa.
}
\description{
Ustawia komunikat systemowy dla konwersacji o podanym ID
}
