% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_system_message}
\alias{set_conversation_system_message}
\title{Sets the system message for the conversation with the given ID}
\usage{
set_conversation_system_message(id, message)
}
\arguments{
\item{id}{Conversation ID.}

\item{message}{New system message (string).}
}
\value{
TRUE if set, FALSE if conversation doesn't exist or value is invalid.
}
\description{
Sets the system message for the conversation with the given ID
}
