% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_temperature}
\alias{set_conversation_temperature}
\title{Ustawia temperaturę dla konwersacji o podanym ID}
\usage{
set_conversation_temperature(id, temperature)
}
\arguments{
\item{id}{ID konwersacji.}

\item{temperature}{Nowa wartość temperatury (liczba 0-1).}
}
\value{
TRUE jeśli ustawiono, FALSE jeśli konwersacja nie istnieje lub wartość jest nieprawidłowa.
}
\description{
Ustawia temperaturę dla konwersacji o podanym ID
}
