% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_menager.R
\name{set_conversation_temperature}
\alias{set_conversation_temperature}
\title{Sets the temperature for the conversation with the given ID}
\usage{
set_conversation_temperature(id, temperature)
}
\arguments{
\item{id}{Conversation ID.}

\item{temperature}{New temperature value (number 0-1).}
}
\value{
TRUE if set, FALSE if conversation doesn't exist or value is invalid.
}
\description{
Sets the temperature for the conversation with the given ID
}
